+++ public/app/features/plugins/built_in_plugins.ts
@@ -44,7 +44,7 @@ import * as textPanel from 'app/plugins/panel/text/module';
 import * as timeseriesPanel from 'app/plugins/panel/timeseries/module';
 import * as stateTimelinePanel from 'app/plugins/panel/state-timeline/module';
 import * as statusHistoryPanel from 'app/plugins/panel/status-history/module';
-import * as marketTrendPanel from 'app/plugins/panel/market-trend/module';
+import * as candlestickPanel from 'app/plugins/panel/candlestick/module';
 import * as graphPanel from 'app/plugins/panel/graph/module';
 import * as xyChartPanel from 'app/plugins/panel/xychart/module';
 import * as dashListPanel from 'app/plugins/panel/dashlist/module';
@@ -100,7 +100,7 @@ const builtInPlugins: any = {
   'app/plugins/panel/timeseries/module': timeseriesPanel,
   'app/plugins/panel/state-timeline/module': stateTimelinePanel,
   'app/plugins/panel/status-history/module': statusHistoryPanel,
-  'app/plugins/panel/market-trend/module': marketTrendPanel,
+  'app/plugins/panel/candlestick/module': candlestickPanel,
   'app/plugins/panel/graph/module': graphPanel,
   'app/plugins/panel/xychart/module': xyChartPanel,
   'app/plugins/panel/geomap/module': geomapPanel,
